requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Redis
redis==5.0.1
hiredis==2.3.2

# LLM Integrations
openai==1.57.0
anthropic==0.39.0

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
httpx==0.26.0
pyyaml==6.0.1

